v1/requirements.txt
PyYAML==6.0.2
streamlit==1.38.0
